Add files via upload
This exercise focuses on path planning for a robot using search algorithms to navigate from a start to a goal position in a known environment. The task involves implementing a greedy uninformed search that explores paths without heuristics and an A algorithm* that integrates a heuristic (Euclidean distance) for efficient pathfinding. The robot moves through an occupancy grid map, considering valid neighboring positions, and the algorithms are evaluated based on path length and the number of explored nodes.